Remove the old `WPMediaUtils` class
