+++ WordPress/src/main/java/org/wordpress/android/ui/prefs/AppPrefs.java
@@ -18,7 +18,7 @@ import org.wordpress.android.ui.comments.CommentsListFragment.CommentStatusCrite
 import org.wordpress.android.ui.reader.utils.ReaderUtils;
 import org.wordpress.android.ui.stats.StatsTimeframe;
 import org.wordpress.android.util.StringUtils;
-import org.wordpress.android.util.WPMediaUtils;
+import org.wordpress.android.util.WordPressMediaUtils;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -587,7 +587,7 @@ public class AppPrefs {
 
     public static int getImageOptimizeWidth() {
         int resizeWidth = getInt(DeletablePrefKey.IMAGE_OPTIMIZE_WIDTH, 0);
-        return resizeWidth == 0 ? WPMediaUtils.OPTIMIZE_IMAGE_MAX_WIDTH : resizeWidth;
+        return resizeWidth == 0 ? WordPressMediaUtils.OPTIMIZE_IMAGE_MAX_WIDTH : resizeWidth;
     }
 
     public static void setImageOptimizeQuality(int quality) {
@@ -596,7 +596,7 @@ public class AppPrefs {
 
     public static int getImageOptimizeQuality() {
         int quality = getInt(DeletablePrefKey.IMAGE_OPTIMIZE_QUALITY, 0);
-        return quality > 1 ? quality : WPMediaUtils.OPTIMIZE_IMAGE_ENCODER_QUALITY;
+        return quality > 1 ? quality : WordPressMediaUtils.OPTIMIZE_IMAGE_ENCODER_QUALITY;
     }
 
     public static boolean isVideoOptimize() {
@@ -613,7 +613,7 @@ public class AppPrefs {
 
     public static int getVideoOptimizeWidth() {
         int resizeWidth =getInt(DeletablePrefKey.VIDEO_OPTIMIZE_WIDTH, 0);
-        return resizeWidth == 0 ? WPMediaUtils.OPTIMIZE_VIDEO_MAX_WIDTH : resizeWidth;
+        return resizeWidth == 0 ? WordPressMediaUtils.OPTIMIZE_VIDEO_MAX_WIDTH : resizeWidth;
     }
 
     public static void setVideoOptimizeQuality(int quality) {
@@ -622,7 +622,7 @@ public class AppPrefs {
 
     public static int getVideoOptimizeQuality() {
         int quality =  getInt(DeletablePrefKey.VIDEO_OPTIMIZE_QUALITY, 0);
-        return quality > 1 ? quality : WPMediaUtils.OPTIMIZE_VIDEO_ENCODER_BITRATE_KB;
+        return quality > 1 ? quality : WordPressMediaUtils.OPTIMIZE_VIDEO_ENCODER_BITRATE_KB;
     }
 
     /*
